Melhorando fluxo de dados do sistema

+++ comunicavet_api/src/main/java/com/projetointegrador/comunicavet/dto/contact/RequestContactDTO.java
@@ -2,7 +2,7 @@ package com.projetointegrador.comunicavet.dto.contact;
 
 // Uso: quando for criar um novo Contato para Clinica; alterar
 public record RequestContactDTO(
-        Byte id,
+        Long id,
         Long clinicId,
         Byte contactTypeId,
         String value
